# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  CreateParcelRequest:
    properties:
      note:
        type: string
      picked_up_date:
        type: string
      received_date:
        type: string
      recipient_name:
        type: string
      room_number:
        type: string
      status:
        $ref: '#/definitions/ParcelStatus'
      tracking_number:
        type: string
    type: object
  CreateParkingSlotRequest:
    properties:
      license_plate:
        type: string
      monthly_fee:
        type: number
      note:
        type: string
      slot_number:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/ParkingStatus'
      tenant_id:
        type: string
      vehicle_type:
        $ref: '#/definitions/VehicleType'
    type: object
  CreatePaymentRequest:
    properties:
      bill_id:
        type: string
      note:
        type: string
      payment_date:
        type: string
      splits:
        items:
          $ref: '#/definitions/PaymentSplitRequest'
        type: array
    type: object
...
  ExpenseReport:
    properties:
      grand_total:
        type: number
      rows:
        items:
          $ref: '#/definitions/ExpenseReportRow'
        type: array
    type: object
  ExpenseReportRow:
    properties:
      category:
        type: string
      count:
        type: integer
      month:
        type: string
      total_amount:
        type: number
    type: object
  IncomeReport:
    properties:
      grand_total:
        type: number
      rows:
        items:
          $ref: '#/definitions/IncomeReportRow'
        type: array
      total_electric:
        type: number
      total_other:
        type: number
      total_rent:
        type: number
      total_water:
        type: number
    type: object
  IncomeReportRow:
    properties:
      bill_count:
        type: integer
      electric_amount:
        type: number
      month:
        type: string
      other_amount:
        type: number
      rent_amount:
        type: number
      total_amount:
        type: number
      water_amount:
        type: number
    type: object
...
  OccupancyReport:
    properties:
      available:
        type: integer
      maintenance:
        type: integer
      occupancy_rate:
        type: number
      occupied:
        type: integer
      rows:
        items:
          $ref: '#/definitions/OccupancyRow'
        type: array
      total:
        type: integer
    type: object
  OccupancyRow:
    properties:
      floor:
        type: integer
      rent_price:
        type: number
      room_number:
        type: string
      status:
        type: string
      tenant_name:
        type: string
      type:
        type: string
    type: object
  Parcel:
    properties:
      created_at:
        type: string
      dormitory_id:
        type: string
      id:
        type: string
      note:
        type: string
      picked_up_date:
        type: string
      received_date:
        type: string
      recipient_name:
        type: string
      room_number:
        type: string
      status:
        $ref: '#/definitions/ParcelStatus'
      tracking_number:
        type: string
      updated_at:
        type: string
    type: object
  ParcelStatus:
    enum:
    - pending
    - picked_up
    type: string
    x-enum-varnames:
    - ParcelStatusPending
    - ParcelStatusPickedUp
  ParkingSlot:
    properties:
      created_at:
        type: string
      dormitory_id:
        type: string
      id:
        type: string
      license_plate:
        type: string
      monthly_fee:
        type: number
      note:
        type: string
      slot_number:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/ParkingStatus'
      tenant_first_name:
        description: joined
        type: string
      tenant_id:
        type: string
      tenant_last_name:
        type: string
      updated_at:
        type: string
      vehicle_type:
        $ref: '#/definitions/VehicleType'
    type: object
  ParkingStatus:
    enum:
    - available
    - occupied
    type: string
    x-enum-varnames:
    - ParkingStatusAvailable
    - ParkingStatusOccupied
  Payment:
    properties:
      amount:
        type: number
      bill_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      method:
        $ref: '#/definitions/PaymentMethod'
      note:
        type: string
      payment_date:
        type: string
      splits:
        items:
          $ref: '#/definitions/PaymentSplit'
        type: array
      updated_at:
        type: string
    type: object
  PaymentDetail:
    properties:
      amount:
        type: number
      bill_id:
        type: string
      bill_total:
        type: number
      billing_month:
        type: string
      created_at:
        type: string
      id:
        type: string
      method:
        $ref: '#/definitions/PaymentMethod'
      note:
        type: string
      payment_date:
        type: string
      room_number:
        type: string
      splits:
        items:
          $ref: '#/definitions/PaymentSplit'
        type: array
      tenant_first_name:
        type: string
      tenant_last_name:
        type: string
      updated_at:
        type: string
    type: object
  PaymentMethod:
    enum:
    - cash
    - transfer
    - qr
    - mixed
    type: string
    x-enum-varnames:
    - PaymentMethodCash
    - PaymentMethodTransfer
    - PaymentMethodQR
    - PaymentMethodMixed
  PaymentSplit:
    properties:
      amount:
        type: number
      id:
        type: string
      method:
        $ref: '#/definitions/PaymentMethod'
      payment_id:
        type: string
    type: object
  PaymentSplitRequest:
    properties:
      amount:
        type: number
      method:
        $ref: '#/definitions/PaymentMethod'
    type: object
...
  UpdateParcelRequest:
    properties:
      note:
        type: string
      picked_up_date:
        type: string
      received_date:
        type: string
      recipient_name:
        type: string
      room_number:
        type: string
      status:
        $ref: '#/definitions/ParcelStatus'
      tracking_number:
        type: string
    type: object
  UpdateParkingSlotRequest:
    properties:
      license_plate:
        type: string
      monthly_fee:
        type: number
      note:
        type: string
      slot_number:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/ParkingStatus'
      tenant_id:
        type: string
      vehicle_type:
        $ref: '#/definitions/VehicleType'
    type: object
  UpdatePaymentRequest:
    properties:
      note:
        type: string
      payment_date:
        type: string
      splits:
        items:
          $ref: '#/definitions/PaymentSplitRequest'
        type: array
    type: object
...
  VehicleType:
    enum:
    - car
    - motorcycle
    type: string
    x-enum-varnames:
    - VehicleTypeCar
    - VehicleTypeMotorcycle
...
info:
  contact: {}
  title: API
paths:
...
  /parcels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Parcel'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายการพัสดุ
      tags:
      - parcels
    post:
      consumes:
      - application/json
      parameters:
      - description: Parcel payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateParcelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Parcel'
      security:
      - ApiKeyAuth: []
      summary: สร้างรายการพัสดุ
      tags:
      - parcels
  /parcels/{id}:
    delete:
      parameters:
      - description: Parcel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบรายการพัสดุ
      tags:
      - parcels
    get:
      parameters:
      - description: Parcel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Parcel'
      security:
      - ApiKeyAuth: []
      summary: ดูข้อมูลพัสดุตาม ID
      tags:
      - parcels
    put:
      consumes:
      - application/json
      parameters:
      - description: Parcel ID
        in: path
        name: id
        required: true
        type: string
      - description: Parcel payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateParcelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Parcel'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขรายการพัสดุ
      tags:
      - parcels
  /parking:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ParkingSlot'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายการที่จอดรถ
      tags:
      - parking
    post:
      consumes:
      - application/json
      parameters:
      - description: Parking slot payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateParkingSlotRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ParkingSlot'
      security:
      - ApiKeyAuth: []
      summary: สร้างที่จอดรถ
      tags:
      - parking
  /parking/{id}:
    delete:
      parameters:
      - description: Parking Slot ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบที่จอดรถ
      tags:
      - parking
    get:
      parameters:
      - description: Parking Slot ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ParkingSlot'
      security:
      - ApiKeyAuth: []
      summary: ดูข้อมูลที่จอดรถตาม ID
      tags:
      - parking
    put:
      consumes:
      - application/json
      parameters:
      - description: Parking Slot ID
        in: path
        name: id
        required: true
        type: string
      - description: Parking slot payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateParkingSlotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ParkingSlot'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขที่จอดรถ
      tags:
      - parking
  /payments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Payment'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายการชำระเงิน
      tags:
      - payments
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Payment'
      security:
      - ApiKeyAuth: []
      summary: สร้างรายการชำระเงิน
      tags:
      - payments
  /payments/{id}:
    delete:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบรายการชำระเงิน
      tags:
      - payments
    get:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PaymentDetail'
      security:
      - ApiKeyAuth: []
      summary: ดูข้อมูลการชำระเงินตาม ID
      tags:
      - payments
    put:
      consumes:
      - application/json
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Payment'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขรายการชำระเงิน
      tags:
      - payments
...
  /reports/expenses:
    get:
      parameters:
      - description: From date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: To date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ExpenseReport'
      security:
      - ApiKeyAuth: []
      summary: รายงานรายจ่าย
      tags:
      - reports
  /reports/income:
    get:
      parameters:
      - description: From date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: To date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/IncomeReport'
      security:
      - ApiKeyAuth: []
      summary: รายงานรายรับ
      tags:
      - reports
  /reports/occupancy:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OccupancyReport'
      security:
      - ApiKeyAuth: []
      summary: รายงานอัตราการเข้าพัก
      tags:
      - reports
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"